kivy/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
docutils==0.17.1
idna==3.2
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
Pillow==8.3.1
pygame==2.0.1
Pygments==2.9.0
PyYAML==5.4.1
requests==2.26.0
urllib3==1.26.6
